Moved marshalling code to own namespace
All decoding related code was moved to Dhall.Marshal.Decode and the
encoding related code to Dhall.Marshal.Encode respectively.

Additionally some minor documentation issues were fixed.
